Add onnxscript 0.4.0 as a dependency
According to the pytorch docs (https://docs.pytorch.org/docs/stable/onnx_export.html),
onnx export depends on onnxscript being installed. In PyTorch 2.9, the export behavior
now defaults to torch.export instead of torch.onnx, which is a codepath that imports
onnxscript eagerly. An alternative is to set dynamo=False on onnx export, but
this is not recommended by PyTorch release notes.

https://github.com/pytorch/pytorch/releases/tag/v2.9.0